+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Documentation: Fix link to GitHub Actions CI configuration.
 - feat: Add tags to logfmt formatter output
+- fix: Ensure time_key is used in logfmt formatter
 
 ## [5.0.0]